# Add simd optimized escape