```
Escaping                time:   [2.1327 us 2.1333 us 2.1340 us]
                        change: [-57.511% -57.410% -57.333%] (p = 0.00 < 0.05)

Big table               time:   [756.68 us 757.40 us 758.59 us]
                        change: [-2.1775% +1.2301% +4.3756%] (p = 0.47 > 0.05)

Teams                   time:   [852.95 ns 853.39 ns 854.04 ns]
                        change: [+0.6336% +0.7563% +0.8888%] (p = 0.00 < 0.05)
```