@botika 2018-12-03
The bench `empty` its remove because (for some strange reason, UFOs?), it gives an error in `(signal: 11, SIGSEGV: invalid memory reference)`. But it passes ALL the tests of each of the libraries and works in beta and nightly. And of course it works correctly out of criterion.